third_party/blink/web_tests/netinfo/gc-unused-listeners-expected.txt

Tests that unused listeners are collected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS callbackObserver.wasCollected is false
PASS callbackObserver.wasCollected is true
PASS successfullyParsed is true

TEST COMPLETE
